project7_requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
openai-whisper==20231117
openai==1.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyAudio==0.2.14
ffmpeg-python==0.2.0
numpy==1.24.3
Werkzeug==3.0.1